Setup global database transaction in tests
This allows each test to have its own isolated database environment where it can start from a blank slate and only insert the data it needs for the test without affecting other tests. This means that a POST call does not affect another GET call due to potentially writing to the database at the same time.