Remove process.exit from listen on error
In theory, I think what would do is, if an uncaught error comes up, it would shut down the back-end. I don't want the back-end to shut down completely if an error occurs, of course. The Sentry integration will come soon, once I set the fundamental rules of how endpoints in general work, but in any case, we should not be exiting on error.