Add --suppress-all to format script
This ensures that any ESLint errors that come up when formatting are suppressed so that rules that would be caught by running the lint script don't block formatting.

Note that a side effect of this is that it generates an eslint-suppressions.json file by default, which helps people on legacy codebases to store existing violations without them blocking in the moment. In our case, however, we don't need this - we just want errors to be suppressed while fixing, while we rely on linting to catch errors instead.