Refactor tests to get rid of special memberSince treatment
Now that the parser functions will convert strings to dates for us, we don't need to create special cases for memberSince anymore. We also no longer need to do all that key deletion - the better practice now is to assert that the API user is at least a subset of factoryUser, and then afterwards, assert the properties that the API user should not have.