Remove validation of POST body in test
The POST endpoint should be the one responsible for validating the POST request body. The front-end is most likely not going to validate the request body itself since that tends to be more of a back-end responsibility. As such, it's more accurate to the actual flow if we don't validate the request body before sending it to the back-end, and just trust that the back-end will validate it in its own way.