Restrict access to import of @neurosongs/prisma-client
The back-end gets everything except the generated types since it can instead rely on @neurosongs/types for all the Zod schemas and types. The front-end cannot access the Prisma client at all, since it's better practice to go through the back-end so it can apply extra validation that cannot be bypassed from the front-end.