Make getPrismaClient a default export
This makes getPrismaClient the top-level import of the connection file, while setPrismaClient is a regular named import. This subtly helps communicate that getPrismaClient is the important one, while setPrismaClient is tucked away a bit more (since we don't want to allow it to be imported anywhere outside the setup files).