Add public redirects
Preparation for when the front-end eventually gets deployed. This is generally needed because React is a single-page app, so we need something to tell it that all paths on our domain should always navigate back to the index.html file, which just loads different content depending on what we want to render.