Add username to public song information
This is useful to have so that we can display both a user's artist name and username on the same song card. Artist names and usernames in Neurosongs are essentially like global names and tags in other social media platforms where the username is a way to uniquely identify a user and must be unique to the platform, and the artist name is their main name and is allowed to not be unique.