Use Loader component
This simplifies pages that do queries by a lot by getting rid of the need to keep providing those if statements to check the loading and error states. Now we can just wrap the parts we want to load with the Loader component, and it will take care of the loading/error displays for us.

Note that we also have to wrap around the Loader component from my components library so that we can use our own default Neurosongs error message throughout, as opposed to the more generalised default in components. This simplifies our usage further so that we don't need to keep passing in the errorComponent every single time, since we'd most likely do that most of the time in Neurosongs.