Separate out router components into their own files
This makes the base router feel tidier, since each resource controls its own router and then the main one can just import it and use that.

Note, however, that for some reason, doing something like <UsersRouter /> to pass the routes does not fully work. I think this is because <UsersRouter /> is interpreted as a component in itself rather than purely as a wrapper around the Route component. The Routes needs to see the Route component directly in order to work, and this is only possible when we actually call UsersRouter() as a raw function rather than a JSX element.