Move RecentSongs and RecentAlbums to songs and albums resource folder
I think they makes more sense there Recent isn't really a resource in itself and the contents of those tabs are more directly linked with the resource. It's a slightly ambiguous case in general whenever we have relations like that, but I think the pattern I'll go for is that for things that aren't really resources (e.g. the songs tab on a Recents page), group it with the resource we're displaying, and for things that are database resources (e.g. UserSongs), group it with the users.