Add a step to ensure lockfile changes get committed
This prevents a rogue package-lock.json change from not making it into a pull request. It can be very easy to forget this in a Turborepo where you can have your terminal set to different parts of the repository to run specific commands/scripts in just the isolated parts, but then you have to remember that the package-lock.json file is at the root and is not managed per workspace.