Ensure that we map through all error codes, including ones coming from apiErrorMap
Previously, we were only mapping through the default error codes. This change ensures that we map through the given ones as well.

Also pretty nice that it involves the removeDuplicates function, one of the first bootcamp problems that I actually struggled with, and now it actually has a use case! The nostalgia...