Add Error boundary
This ensures that, if we ever get a case where the page crashes entirely, it renders an actual message rather than just a blank screen. This is helpful as it ensures that the crashed page maintains the same theme, avoiding dark mode users suddenly get blinded by a big white screen, as well as also surfacing the exact error that occurred.