Uninstall ESLint-related dependencies from front-end
Since ESLint and my ESLint plugin are dependencies used throughout this tech stack, they should only ever be installed at the root so that all workspaces can share the same version of the plugin and ESLint. The front-end only really had it because of the Vite template, but these have since been uninstalled.