Separate out the local packages scripts to refresh-local-package/use-local-package and refresh-live-package/use-live-package
The refresh variants only swap out the live package for local or local for live. The use variants do that and also starts the dev server. This makes iterating on changes a little easier as you can just run one command to swap out the package and start the dev server, which is common to do in my workflow.

Note that I have also included all other packages I have as well for the sake of consistency. In practice, I'd rarely use local utility as utility functions are more lightweight and easily testable with its own test suite, but it's still good to have around anyway.