Rename refresh scripts to prepare
This feels clearer as the main concern of those scripts is to prepare the package. Refreshing is a secondary concern.

Note that the dev-refresh and dev-after-refresh scripts are still named as such, though. This is because the concern of dev-refresh is genuinely to refresh the cache rather than prepare something. As such, that makes sense as refresh, but the other scripts have been renamed to prepare.