Deal with linting errors as a result of updating the ESLint plugin
This includes installing eslint-plugin-perfectionist and applying the sorting rules. Note that I had to delete parseDate as it was not being used, and add JSDoc comments to paginate. This is because the utility package here also uses the utility config, which also expects JSDoc comments. Whether that's super useful here or not, I don't know, but it might be good to keep around anyway. For types, though, it has been disabled as it is not that useful there.