Fix parseEnv error in back-end
If NODE_ENV is undefined, it doesn't necessarily fall back to development, hence a default must be provided. I missed it out initially because the tests run in a test environment and not development, hence we never got the chance to test out pnpm run dev in CI and watch it fail.